requirements.txt
streamlit
pydash
streamlit-tags
Pillow
streamlit-quill
